Added VTP support for VTK.
Now you can load vtp files (can be binary, ascii and compressed).
Also added a example related to it. It depends on the Zlib.js inflate.
It can be optimized a little more.

I havent tested on large variety of files, but for smaller ones, it should work.
It does not support Int64 numbers yet, so any model that has such numbers might not load correctly.
There is still a problem of converting the binary compressed string to the original data if the string is not in this format XXXX==XXXX. But it works for most cases.